% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NaviCell_RC.R
\docType{class}
\name{NaviCell-class}
\alias{NaviCell}
\alias{NaviCell-class}
\title{NaviCell reference class}
\description{
NaviCell reference class
}
\section{Methods}{

\describe{
\item{\code{attachLastSession(...)}}{Attach NaviCell handle to the last existing NaviCell Web Service session.}

\item{\code{attachSession(session_id)}}{Attach a NaviCell server session ID.}

\item{\code{formatJson(list_param)}}{Format list of parameters to NaviCell server compatible JSON format (internal utility).}

\item{\code{formatResponse(response)}}{Format response obtained from the RCurl 'postForm' command (internal utility).}

\item{\code{generateSessionId(...)}}{Generate a session ID.}

\item{\code{incMessageId(...)}}{Increase message ID counter.}

\item{\code{launchBrowser(...)}}{Launch client browser and points to the default NaviCell map.}

\item{\code{listSessions(...)}}{List all NaviCell server sessions.}

\item{\code{makeData(json_string)}}{Create NaviCell server command string from a list of parameters (internal utility).}

\item{\code{setZoom(module, zoom_level)}}{Set a given zoom level on associated NaviCell map in browser.}
}}

